Finalize Option G as the MCP settings screen design
Remove all exploration infrastructure (variations A-F, ExplorationsHelper,
localStorage variation switching) and keep only Option G for both the
modern dashboard and legacy Calypso interfaces.

- Rewrite dashboard index.tsx and legacy main.jsx with Option G two-card layout
- Fix routes to point back to original files
- Remove ExplorationsHelper from app root
- Clean up variation conditionals in sub-pages (ai-assistant, sites, setup, tools-category)
- Delete unused files (index-explore.tsx, main-explore.jsx, tools.jsx, explorations-helper/)
- Fix list styles for dashboard summary-button-list in legacy Calypso
