+++ client/me/mcp/index.js
@@ -5,17 +5,15 @@ import {
 	maybeRedirectToMultiSiteDashboard,
 } from 'calypso/controller';
 import { setupPreferences } from 'calypso/controller/preferences';
-// EXPLORE: temporarily using mcpExplore (original: mcp)
-import { sidebar, mcpExplore } from 'calypso/me/controller';
+import { sidebar, mcp } from 'calypso/me/controller';
 
 export default function () {
-	// EXPLORE: temporarily pointing to mcpExplore (original: mcp)
 	page(
 		'/me/mcp',
 		setupPreferences,
 		maybeRedirectToMultiSiteDashboard( '/me/mcp' ),
 		sidebar,
-		mcpExplore,
+		mcp,
 		makeLayout,
 		clientRender
 	);
