+++ client/me/mcp/setup.jsx
@@ -16,13 +16,11 @@ import { copy, check, error } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
 import { useState, useEffect } from 'react';
 import { useDispatch } from 'react-redux';
-import CardHeading from 'calypso/components/card-heading';
 import DocumentHead from 'calypso/components/data/document-head';
 import HeaderCake from 'calypso/components/header-cake';
 import InlineSupportLink from 'calypso/components/inline-support-link';
 import Main from 'calypso/components/main';
 import NavigationHeader from 'calypso/components/navigation-header';
-import LegacySectionHeader from 'calypso/components/section-header';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
 import twoStepAuthorization from 'calypso/lib/two-step-authorization';
 import ReauthRequired from 'calypso/me/reauth-required';
@@ -31,8 +29,6 @@ import { SectionHeader } from '../../dashboard/components/section-header';
 import { hasEnabledAccountTools } from './utils';
 
 import './style.scss';
-
-const EXPLORATIONS_STORAGE_KEY = 'mcp-explore-variation';
 
 function McpSetupComponent( { path } ) {
 	const translate = useTranslate();
@@ -204,9 +200,6 @@ function McpSetupComponent( { path } ) {
 		}
 	};
 
-	const variation = localStorage.getItem( EXPLORATIONS_STORAGE_KEY );
-	const isSquareCorners = variation === 'F' || variation === 'G';
-
 	// Handle error states only - allow loading to continue so reauth can show
 	if ( userSettingsError ) {
 		return null;
@@ -238,64 +231,20 @@ function McpSetupComponent( { path } ) {
 	const hasEnabledTools = hasEnabledAccountTools( userSettings );
 
 	if ( ! hasEnabledTools ) {
-		const isLegacyEmpty = variation === 'F';
 		return renderLayout(
-			isLegacyEmpty ? (
-				<>
-					<LegacySectionHeader label={ translate( 'Enable AI access first' ) } />
-					<Card isRounded={ false }>
-						<CardBody>
-							<VStack spacing={ 4 }>
-								<p style={ { color: '#646970', margin: 0, fontSize: '14px' } }>
-									{ translate( 'You need to enable AI access before connecting your assistant.' ) }
-								</p>
-								<Button variant="primary" href="/me/mcp" style={ { alignSelf: 'flex-start' } }>
-									{ translate( 'Go to AI and MCP' ) }
-								</Button>
-							</VStack>
-						</CardBody>
-					</Card>
-				</>
-			) : (
-				<Card
-					isRounded={ ! isSquareCorners }
-					style={ isSquareCorners ? { borderRadius: 0 } : undefined }
-				>
-					<CardBody>
-						<VStack spacing={ 4 }>
-							<SectionHeader
-								level={ 3 }
-								title={
-									isSquareCorners
-										? translate( 'Enable AI access first' )
-										: translate( 'Setup Required' )
-								}
-							/>
-							<Text
-								as="p"
-								color={ isSquareCorners ? '#646970' : undefined }
-								size={ isSquareCorners ? undefined : 'medium' }
-							>
-								{ isSquareCorners
-									? translate( 'You need to enable AI access before connecting your assistant.' )
-									: translate( 'No MCP tools are currently enabled for your account.' ) }
-							</Text>
-							{ ! isSquareCorners && (
-								<Text as="p" size="medium">
-									{ translate(
-										'MCP tools define what actions and data your MCP client can access on your account. You need to enable at least one tool in the main MCP settings before configuring your client.'
-									) }
-								</Text>
-							) }
-							<Button variant="primary" href="/me/mcp" style={ { alignSelf: 'flex-start' } }>
-								{ isSquareCorners
-									? translate( 'Go to AI and MCP' )
-									: translate( 'Go to MCP Settings' ) }
-							</Button>
-						</VStack>
-					</CardBody>
-				</Card>
-			)
+			<Card isRounded={ false } style={ { borderRadius: 0 } }>
+				<CardBody>
+					<VStack spacing={ 4 }>
+						<SectionHeader level={ 3 } title={ translate( 'Enable AI access first' ) } />
+						<Text as="p" color="#646970">
+							{ translate( 'You need to enable AI access before connecting your assistant.' ) }
+						</Text>
+						<Button variant="primary" href="/me/mcp" style={ { alignSelf: 'flex-start' } }>
+							{ translate( 'Go to AI and MCP' ) }
+						</Button>
+					</VStack>
+				</CardBody>
+			</Card>
 		);
 	}
 
@@ -411,486 +360,90 @@ function McpSetupComponent( { path } ) {
 		}
 	};
 
-	if ( isSquareCorners ) {
-		const quickSetup = renderQuickSetupG();
-		const isLegacyStyle = variation === 'F';
-
-		return renderLayout(
-			<>
-				<HeaderCake backText={ translate( 'Back' ) } backHref="/me/mcp">
-					{ translate( 'Connect AI assistant' ) }
-				</HeaderCake>
-				{ isLegacyStyle ? (
-					<>
-						{ /* Client picker */ }
-						<LegacySectionHeader label={ translate( 'AI assistant' ) } />
-						<Card isRounded={ false }>
-							<CardBody>
-								<SelectControl
-									__next40pxDefaultSize
-									__nextHasNoMarginBottom
-									label={ translate( 'Choose your AI assistant' ) }
-									value={ selectedMcpClient }
-									options={ mcpClientOptions }
-									onChange={ setSelectedMcpClient }
-								/>
-							</CardBody>
-						</Card>
-
-						{ /* Quick setup (client-specific) */ }
-						{ quickSetup && (
-							<>
-								<LegacySectionHeader
-									label={ translate( 'Quick setup' ) }
-									className="mcp__section-header"
-								/>
-								<Card isRounded={ false }>
-									<CardBody>{ quickSetup }</CardBody>
-								</Card>
-							</>
-						) }
-
-						{ /* Manual setup — JSON config */ }
-						<LegacySectionHeader
-							label={ translate( 'Manual setup' ) }
-							className="mcp__section-header"
-						>
-							<Button
-								icon={ getCopyIcon() }
-								variant="tertiary"
-								iconSize={ 16 }
-								size="small"
-								style={ {
-									minWidth: '24px',
-									height: '24px',
-									padding: 0,
-									color: copyStatus === 'error' ? 'var(--color-error)' : undefined,
-								} }
-								onClick={ copyToClipboard }
-								aria-label={ translate( 'Copy configuration to clipboard' ) }
-							/>
-						</LegacySectionHeader>
-						<Card isRounded={ false }>
-							<CardBody>
-								<VStack spacing={ 4 }>
-									<p style={ { color: '#646970', margin: 0, fontSize: '14px' } }>
-										{ translate( "Copy this configuration into your client's MCP settings." ) }
-									</p>
-									<TextareaControl
-										__nextHasNoMarginBottom
-										value={ JSON.stringify( generateMcpConfig( selectedMcpClient ), null, 2 ) }
-										onChange={ () => {} }
-										readOnly
-										style={ { minHeight: '140px' } }
-									/>
-									{ clientDocumentation[ selectedMcpClient ] && (
-										<ExternalLink
-											href={ clientDocumentation[ selectedMcpClient ] }
-											style={ { fontSize: '13px' } }
-										>
-											{ selectedMcpClient === 'default'
-												? translate( 'View setup instructions' )
-												: sprintf( translate( '%s documentation' ), clientLabel ) }
-										</ExternalLink>
-									) }
-								</VStack>
-							</CardBody>
-						</Card>
-					</>
-				) : (
-					<VStack spacing={ 6 }>
-						{ /* Card 1: Client picker */ }
-						<Card isRounded={ false } style={ { borderRadius: 0 } }>
-							<CardBody>
-								<SelectControl
-									__next40pxDefaultSize
-									__nextHasNoMarginBottom
-									label={ translate( 'Choose your AI assistant' ) }
-									value={ selectedMcpClient }
-									options={ mcpClientOptions }
-									onChange={ setSelectedMcpClient }
-								/>
-							</CardBody>
-						</Card>
-
-						{ /* Card 2: Quick setup (client-specific) */ }
-						{ quickSetup && (
-							<Card isRounded={ false } style={ { borderRadius: 0 } }>
-								<CardBody>
-									<VStack spacing={ 4 }>
-										<SectionHeader level={ 3 } title={ translate( 'Quick setup' ) } />
-										{ quickSetup }
-									</VStack>
-								</CardBody>
-							</Card>
-						) }
-
-						{ /* Card 3: Manual setup — JSON config */ }
-						<Card isRounded={ false } style={ { borderRadius: 0 } }>
-							<CardBody>
-								<VStack spacing={ 4 }>
-									<div
-										style={ {
-											display: 'flex',
-											justifyContent: 'space-between',
-											alignItems: 'center',
-										} }
-									>
-										<SectionHeader
-											level={ 3 }
-											title={ translate( 'Manual setup' ) }
-											description={ translate(
-												"Copy this configuration into your client's MCP settings."
-											) }
-										/>
-										<Button
-											icon={ getCopyIcon() }
-											variant="tertiary"
-											iconSize={ 20 }
-											style={ {
-												color: copyStatus === 'error' ? 'var(--color-error)' : undefined,
-											} }
-											onClick={ copyToClipboard }
-											aria-label={ translate( 'Copy configuration to clipboard' ) }
-										/>
-									</div>
-									<TextareaControl
-										__nextHasNoMarginBottom
-										value={ JSON.stringify( generateMcpConfig( selectedMcpClient ), null, 2 ) }
-										onChange={ () => {} }
-										readOnly
-										style={ { minHeight: '140px' } }
-									/>
-									{ clientDocumentation[ selectedMcpClient ] && (
-										<ExternalLink
-											href={ clientDocumentation[ selectedMcpClient ] }
-											style={ { fontSize: '13px' } }
-										>
-											{ selectedMcpClient === 'default'
-												? translate( 'View setup instructions' )
-												: sprintf( translate( '%s documentation' ), clientLabel ) }
-										</ExternalLink>
-									) }
-								</VStack>
-							</CardBody>
-						</Card>
-					</VStack>
-				) }
-			</>
-		);
-	}
+	const quickSetup = renderQuickSetupG();
 
 	return renderLayout(
 		<>
 			<HeaderCake backText={ translate( 'Back' ) } backHref="/me/mcp">
-				{ translate( 'WordPress.com MCP Setup' ) }
+				{ translate( 'Connect AI assistant' ) }
 			</HeaderCake>
-
-			<Card>
-				<CardBody>
-					<VStack spacing={ 6 }>
+			<VStack spacing={ 6 }>
+				{ /* Card 1: Client picker */ }
+				<Card isRounded={ false } style={ { borderRadius: 0 } }>
+					<CardBody>
 						<SelectControl
 							__next40pxDefaultSize
 							__nextHasNoMarginBottom
-							label={ translate( 'MCP Client' ) }
+							label={ translate( 'Choose your AI assistant' ) }
 							value={ selectedMcpClient }
 							options={ mcpClientOptions }
 							onChange={ setSelectedMcpClient }
-							help={ translate(
-								'Choose your MCP client to get the correct configuration format.'
-							) }
 						/>
+					</CardBody>
+				</Card>
 
-						{ /* Quick Setup for Claude */ }
-						{ selectedMcpClient === 'claude' && (
-							<VStack spacing={ 4 } style={ { borderBottom: '1px solid rgba(0, 0, 0, 0.1)' } }>
-								<CardHeading tagName="h1" size={ 16 } isBold>
-									{ translate( 'Quick Setup' ) }
-								</CardHeading>
-								<Text as="p" size="medium">
-									{ translate( 'For Claude users, connect WordPress.com from Claude Connectors.' ) }
-								</Text>
-								<Text as="p" size="medium">
-									{ translate( 'Installation steps:' ) }
-								</Text>
-								<ol>
-									<li>
-										<Text as="p" size="medium">
-											{ createInterpolateElement(
-												/* translators: %s is the link to the Claude settings page */
-												translate( 'Open <ClaudeSettings/>.' ),
-												{
-													ClaudeSettings: (
-														<ExternalLink href="https://claude.ai/settings/connectors">
-															{ translate( 'Claude settings' ) }
-														</ExternalLink>
-													),
-												}
-											) }
-										</Text>
-									</li>
-									<li>
-										<Text as="p" size="medium">
-											{ translate( 'Click the "Browse connectors" button.' ) }
-										</Text>
-									</li>
-									<li>
-										<Text as="p" size="medium">
-											{ translate( 'Search for WordPress.com.' ) }
-										</Text>
-									</li>
-									<li>
-										<Text as="p" size="medium">
-											{ translate( 'Select WordPress.com and follow the prompts to connect.' ) }
-										</Text>
-									</li>
-								</ol>
+				{ /* Card 2: Quick setup (client-specific) */ }
+				{ quickSetup && (
+					<Card isRounded={ false } style={ { borderRadius: 0 } }>
+						<CardBody>
+							<VStack spacing={ 4 }>
+								<SectionHeader level={ 3 } title={ translate( 'Quick setup' ) } />
+								{ quickSetup }
 							</VStack>
-						) }
+						</CardBody>
+					</Card>
+				) }
 
-						{ /* Quick Setup for Claude Code */ }
-						{ selectedMcpClient === 'claude-code' && (
-							<VStack
-								spacing={ 4 }
-								style={ { borderBottom: '1px solid rgba(0, 0, 0, 0.1)', paddingBottom: '24px' } }
-							>
-								<CardHeading tagName="h1" size={ 16 } isBold>
-									{ translate( 'Quick Setup' ) }
-								</CardHeading>
-								<Text as="p" size="medium">
-									{ translate(
-										'Claude Code uses a different config format with type: "http". Use the CLI or copy the configuration below.'
-									) }
-								</Text>
-								<Text as="p" size="medium">
-									{ translate( 'Installation steps:' ) }
-								</Text>
-								<ol>
-									<li>
-										<Text as="p" size="medium">
-											{ createInterpolateElement(
-												translate( 'Run this command in your terminal: <code>%s</code>' ).replace(
-													'%s',
-													'claude mcp add --transport http wpcom-mcp https://public-api.wordpress.com/wpcom/v2/mcp/v1'
-												),
-												{
-													code: (
-														<code
-															key="claude-code-cmd"
-															style={ {
-																backgroundColor: '#f0f0f1',
-																padding: '2px 6px',
-																borderRadius: '3px',
-																fontFamily: 'monospace',
-																fontSize: '13px',
-															} }
-														>
-															claude mcp add --transport http wpcom-mcp
-															https://public-api.wordpress.com/wpcom/v2/mcp/v1
-														</code>
-													),
-												}
-											) }
-										</Text>
-									</li>
-									<li>
-										<Text as="p" size="medium">
-											{ createInterpolateElement(
-												translate(
-													'Or copy the configuration below and add it to your <mcpJson/> or <claudeJson/> file.'
-												),
-												{
-													mcpJson: (
-														<code
-															key="mcp-json"
-															style={ {
-																backgroundColor: '#f0f0f1',
-																padding: '2px 6px',
-																borderRadius: '3px',
-																fontFamily: 'monospace',
-																fontSize: '13px',
-															} }
-														>
-															.mcp.json
-														</code>
-													),
-													claudeJson: (
-														<code
-															key="claude-json"
-															style={ {
-																backgroundColor: '#f0f0f1',
-																padding: '2px 6px',
-																borderRadius: '3px',
-																fontFamily: 'monospace',
-																fontSize: '13px',
-															} }
-														>
-															~/.claude.json
-														</code>
-													),
-												}
-											) }
-										</Text>
-									</li>
-									<li>
-										<Text as="p" size="medium">
-											{ createInterpolateElement(
-												translate(
-													'In Claude Code, run <code/> to authenticate with your WordPress.com account.'
-												),
-												{
-													code: (
-														<code
-															key="mcp-cmd"
-															style={ {
-																backgroundColor: '#f0f0f1',
-																padding: '2px 6px',
-																borderRadius: '3px',
-																fontFamily: 'monospace',
-																fontSize: '13px',
-															} }
-														>
-															/mcp
-														</code>
-													),
-												}
-											) }
-										</Text>
-									</li>
-								</ol>
-							</VStack>
-						) }
-
-						{ /* Quick Setup for Cursor */ }
-						{ selectedMcpClient === 'cursor' && (
-							<VStack
-								spacing={ 4 }
-								style={ { borderBottom: '1px solid rgba(0, 0, 0, 0.1)', paddingBottom: '24px' } }
-							>
-								<CardHeading tagName="h1" size={ 16 } isBold>
-									{ translate( 'Quick Setup' ) }
-								</CardHeading>
-								<Text as="p" size="medium">
-									{ translate(
-										'For Cursor users, use the one-click install to add the WordPress.com MCP app.'
-									) }
-								</Text>
-								<Button
-									variant="primary"
-									href="cursor://anysphere.cursor-deeplink/mcp/install?name=WordPress.com&config=eyJjb21tYW5kIjoibnB4IC15IG1jcC1yZW1vdGUgaHR0cHM6Ly9wdWJsaWMtYXBpLndvcmRwcmVzcy5jb20vd3Bjb20vdjIvbWNwL3YxIn0%3D"
-									target="_blank"
-									style={ { alignSelf: 'flex-start' } }
-								>
-									{ translate( 'Install in Cursor' ) }
-								</Button>
-							</VStack>
-						) }
-
+				{ /* Card 3: Manual setup — JSON config */ }
+				<Card isRounded={ false } style={ { borderRadius: 0 } }>
+					<CardBody>
 						<VStack spacing={ 4 }>
-							<CardHeading tagName="h1" size={ 16 } isBold>
-								{ translate( 'Manual Setup' ) }
-							</CardHeading>
-
-							<VStack spacing={ 3 }>
-								<div
-									style={ {
-										display: 'flex',
-										justifyContent: 'space-between',
-										alignItems: 'center',
-									} }
-								>
-									{ clientDocumentation[ selectedMcpClient ] && (
-										<ExternalLink
-											href={ clientDocumentation[ selectedMcpClient ] }
-											target="_blank"
-											style={ { fontSize: 'inherit' } }
-										>
-											{ /* translators: %s is the name of the MCP client */ }
-											{ selectedMcpClient === 'default'
-												? translate( 'View setup instructions for other MCP client' )
-												: sprintf(
-														translate( 'View setup instructions for %s' ),
-														mcpClientOptions.find( ( opt ) => opt.value === selectedMcpClient )
-															?.label
-												  ) }
-										</ExternalLink>
+							<div
+								style={ {
+									display: 'flex',
+									justifyContent: 'space-between',
+									alignItems: 'center',
+								} }
+							>
+								<SectionHeader
+									level={ 3 }
+									title={ translate( 'Manual setup' ) }
+									description={ translate(
+										"Copy this configuration into your client's MCP settings."
 									) }
-									<Button
-										icon={ getCopyIcon() }
-										variant="tertiary"
-										size="small"
-										style={ {
-											color: copyStatus === 'error' ? 'var(--color-error)' : undefined,
-										} }
-										onClick={ copyToClipboard }
-										aria-label={ translate( 'Copy configuration to clipboard' ) }
-									/>
-								</div>
-								<TextareaControl
-									__nextHasNoMarginBottom
-									value={ JSON.stringify( generateMcpConfig( selectedMcpClient ), null, 2 ) }
-									onChange={ () => {} } // Required prop for read-only textarea
-									readOnly
-									help={ translate(
-										"Copy this configuration and paste it into your MCP client's settings."
-									) }
-									style={ { minHeight: '160px' } }
 								/>
-							</VStack>
+								<Button
+									icon={ getCopyIcon() }
+									variant="tertiary"
+									iconSize={ 20 }
+									style={ {
+										color: copyStatus === 'error' ? 'var(--color-error)' : undefined,
+									} }
+									onClick={ copyToClipboard }
+									aria-label={ translate( 'Copy configuration to clipboard' ) }
+								/>
+							</div>
+							<TextareaControl
+								__nextHasNoMarginBottom
+								value={ JSON.stringify( generateMcpConfig( selectedMcpClient ), null, 2 ) }
+								onChange={ () => {} }
+								readOnly
+								style={ { minHeight: '140px' } }
+							/>
+							{ clientDocumentation[ selectedMcpClient ] && (
+								<ExternalLink
+									href={ clientDocumentation[ selectedMcpClient ] }
+									style={ { fontSize: '13px' } }
+								>
+									{ selectedMcpClient === 'default'
+										? translate( 'View setup instructions' )
+										: sprintf( translate( '%s documentation' ), clientLabel ) }
+								</ExternalLink>
+							) }
 						</VStack>
-					</VStack>
-					<VStack spacing={ 3 }>
-						<ul style={ { listStyle: 'none', padding: '0', margin: '0' } }>
-							<li>
-								{ createInterpolateElement(
-									translate(
-										'<code>%s</code> is a unique identifier for this WordPress.com account connection'
-									).replace( '%s', serverName ),
-									{
-										code: (
-											<code
-												key="server-name"
-												style={ {
-													backgroundColor: '#f0f0f1',
-													padding: '2px 6px',
-													borderRadius: '3px',
-													fontFamily: 'monospace',
-													fontSize: '13px',
-												} }
-											>
-												{ serverName }
-											</code>
-										),
-									}
-								) }
-							</li>
-							<li>
-								{ createInterpolateElement(
-									translate(
-										'<code>%s</code> is the official WordPress.com MCP server endpoint'
-									).replace( '%s', 'https://public-api.wordpress.com/wpcom/v2/mcp/v1' ),
-									{
-										code: (
-											<code
-												key="package-name"
-												style={ {
-													backgroundColor: '#f0f0f1',
-													padding: '2px 6px',
-													borderRadius: '3px',
-													fontFamily: 'monospace',
-													fontSize: '13px',
-												} }
-											>
-												https://public-api.wordpress.com/wpcom/v2/mcp/v1
-											</code>
-										),
-									}
-								) }
-							</li>
-						</ul>
-					</VStack>
-				</CardBody>
-			</Card>
+					</CardBody>
+				</Card>
+			</VStack>
 		</>
 	);
 }
